requirements.txt
asgiref==3.8.1
Django==5.0.9
django-types==0.19.1
djangorestframework==3.15.2
Pygments==2.18.0
sqlparse==0.5.2
types-psycopg2==2.9.21.20241019
typing_extensions==4.12.2
